backend/requirements.txt
fastapi
uvicorn[standard]
python-decouple
python-multipart
pytest
httpx
pytest-asyncio
anyio
scikit-learn
PyMuPDF
google-generativeai